Several Maintenance Changes (see desc.)
· VS Files added to gitignore (the ones lacking)
· Unnecessary token-paste operator was removed in the event.h (event class type)
· Layer wasn't calling OnDetach() in its destructor, now it does
· Now there won't be issues when trying to update submodules (imgui, spdlog...), since we have specified the current branch we are working with in .gitmodules
· Platform detection was automated in Core.h and deleted from premake, this way we have a way to detect the current platform and specify platform defines
· Solved a couple of cast-warnings
· Shaders now delete when detaching (in its compilation)
· Bat file won't close if suceed, so we can see the result

+++ KaimosEngine/src/Core/Application.cpp
@@ -40,7 +40,7 @@ namespace Kaimos {
 
 		while (m_Running)
 		{
-			float time = glfwGetTime();//QueryPerformanceFrequency(); QueryPerformanceCounter(); // Platform::GetTime() !!!!!!
+			float time = (float)glfwGetTime();//QueryPerformanceFrequency(); QueryPerformanceCounter(); // Platform::GetTime() !!!!!!
 			Timestep timestep = time - m_LastFrameTime; // How long this frame is (dt, current time vs last frame time)
 			m_LastFrameTime = time;
 
@@ -103,7 +103,7 @@ namespace Kaimos {
 
 	bool Application::OnWindowResize(WindowResizeEvent& e)
 	{
-		float w = e.GetWidth(), h = e.GetHeight();
+		uint w = e.GetWidth(), h = e.GetHeight();
 
 		if (w == 0 || h == 0)
 		{
